+++ spec/jobs/donations/update_pool_balance_job_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Donations::GenerateBalanceHistoryJob, type: :job do
+RSpec.describe Donations::UpdatePoolBalanceJob, type: :job do
   describe '#perform' do
     subject(:perform_job) { described_class.perform_now(pool:) }
 
@@ -10,13 +10,13 @@ RSpec.describe Donations::GenerateBalanceHistoryJob, type: :job do
 
     before do
       allow(service).to receive(:new).with(pool:).and_return(service_mock)
-      allow(service_mock).to receive(:add_balance_history)
+      allow(service_mock).to receive(:update_balance)
       perform_job
     end
 
     it 'calls the service with right params' do
       expect(service).to have_received(:new).with(pool:)
-      expect(service_mock).to have_received(:add_balance_history)
+      expect(service_mock).to have_received(:update_balance)
     end
   end
 end
